Refined simulated spiral tests
gal0-4 provided by Chien Peng

Gaussian noise with nsigma added to files.

Fitting gal4 with a two component fit and trying to understand the
results by studying the individual model components, it became
apparent that the spiral arm component included a significant central
component which was confusing the interpretation of the arm component
as additional 'fluff'.

Adapted gal4 to create gal5, which includes an inner trunctation so
that the third component corresponds only to spiral arms.

Adapted gal5 to create gal6, which has a star-formation region in one of
the arms.

Adapted gal6 to create gal7, with slightly asymmetric spiral arms.

Two-sersic feedmes included.


Former-commit-id: 9751929df0e88fd373d45ca56e3512edae28f757